Updated dependencies and refactored code.
* Updated spring cloud streams to 1.1.0.RELEASE
* Updated cloud dependencies parent to 1.2.1.RELEASE
* Updated samples to 1.4.1
* Updated Tests to take advantage of 1.4 boot test annotations
* Removed @Bindings as they are deprecated
* Updated BatchApplication test to pass.  Boot no longer logs environment values.  So count was reduced from 3 to 1

resolves #218, #221

renamed spring-cloud-task-starter to spring-cloud-starter-task
