+++ pom.xml
@@ -125,16 +125,16 @@
 		<module>spring-cloud-task-docs</module>
 		<module>spring-cloud-task-batch</module>
 		<module>spring-cloud-task-stream</module>
-		<module>spring-cloud-task-starter</module>
+		<module>spring-cloud-starter-task</module>
 		<module>spring-cloud-task-samples</module>
 		<module>spring-cloud-task-integration-tests</module>
 	</modules>
 
 	<properties>
-		<spring-cloud-stream.version>1.0.2.RELEASE</spring-cloud-stream.version>
+		<spring-cloud-stream.version>1.1.0.RELEASE</spring-cloud-stream.version>
 		<spring-cloud-deployer-spi.version>1.0.2.RELEASE</spring-cloud-deployer-spi.version>
 		<spring-cloud-deployer-local.version>1.0.2.RELEASE</spring-cloud-deployer-local.version>
-		<spring-cloud-stream-binder-rabbit.version>1.0.2.RELEASE</spring-cloud-stream-binder-rabbit.version>
+		<spring-cloud-stream-binder-rabbit.version>1.1.0.RELEASE</spring-cloud-stream-binder-rabbit.version>
 		<spring-cloud-deployer-resource-support.version>1.0.2.RELEASE</spring-cloud-deployer-resource-support.version>
 		<spring-cloud-deployer-resource-maven.version>1.0.2.RELEASE</spring-cloud-deployer-resource-maven.version>
 		<spring-batch.version>3.0.7.RELEASE</spring-batch.version>
