+++ spring-cloud-starter-task/pom.xml
@@ -9,7 +9,7 @@
 		<version>1.1.0.BUILD-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>spring-cloud-task-starter</artifactId>
+	<artifactId>spring-cloud-starter-task</artifactId>
 	<packaging>jar</packaging>
 	<name>Spring Cloud Task Starter</name>
 	<description>Spring Boot starter for Spring Cloud Task</description>
