+++ spring-cloud-task-dependencies/pom.xml
@@ -10,7 +10,7 @@
 	<parent>
 		<artifactId>spring-cloud-dependencies-parent</artifactId>
 		<groupId>org.springframework.cloud</groupId>
-		<version>1.1.1.RELEASE</version>
+		<version>1.2.1.RELEASE</version>
         <relativePath />
 	</parent>
 
@@ -18,7 +18,7 @@
 		<dependencies>
 			<dependency>
 				<groupId>org.springframework.cloud</groupId>
-				<artifactId>spring-cloud-task-starter</artifactId>
+				<artifactId>spring-cloud-starter-task</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
