+++ spring-cloud-task-samples/tasksink/pom.xml
@@ -13,7 +13,7 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>1.4.0.RELEASE</version>
+		<version>1.4.1.RELEASE</version>
 		<relativePath/>
 	</parent>
 
@@ -46,18 +46,18 @@
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-task-starter</artifactId>
+			<artifactId>spring-cloud-starter-task</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-starter-stream-rabbit</artifactId>
-			<version>1.0.2.RELEASE</version>
+			<version>1.1.0.RELEASE</version>
 			<scope>compile</scope>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-stream-test-support</artifactId>
-			<version>1.0.2.RELEASE</version>
+			<version>1.1.0.RELEASE</version>
 			<scope>test</scope>
 		</dependency>
 		<dependency>
